Remove orphaned operation columns on update
Because the architecture of Roundup re-create Operation columns when the
underlying database view has materialized we need to remove old
columns associated with that operation when these are updated. Note
that this commit adds some technical debit, columns hidden from an
operation via the exclusion mechanisms will also be deleted. But
I have know that the exclusion mechanism needs to be udpated in order
to also support column deletion.