Fix issue with orphaned alerts
This was an interesting bug to fix, so if you change a pack to stack, it would create orphaned alerts, which are alerts associated with an object that are no longer applicable. The stack alerts were still associated with the object when it was convered to a Pack. So we just collect the orphaned alerts and delete them.

Closes #112